pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "deadlymeagerdiscussion"
version = "0.1.0"
[tool.poetry.dependencies]
airtable-python-wrapper = "^0.13.0"
python = "^3.7"
flask = "^1.1"
flask-simple-geoip = "^0.1.1"
simplegeoip = "^0.2.0"
geopy = "^1.21"
translate = "^3.5"
[tool.poetry.dev-dependencies]
